+++ BigWigs_WrathOfTheLichKing/Ulduar/XT-002.lua
@@ -5,7 +5,7 @@
 local mod, CL = BigWigs:NewBoss("XT-002 Deconstructor", 603, 1640)
 if not mod then return end
 mod:RegisterEnableMob(33293)
-mod:SetEncounterID(1142)
+mod:SetEncounterID(mod:Classic() and 747 or 1142)
 -- mod:SetRespawnTime(0) -- resets, doesn't respawn
 
 --------------------------------------------------------------------------------
@@ -82,7 +82,7 @@ end
 function mod:GravityBomb(args)
 	if self:Me(args.destGUID) then
 		self:OpenProximity(64234, 10)
-		self:Say(64234)
+		self:Say(64234, nil, nil, "Gravity Bomb")
 		self:SayCountdown(64234, 9)
 	end
 	self:TargetMessageOld(64234, args.destName, "red", "alert")
@@ -102,7 +102,7 @@ end
 function mod:SearingLight(args)
 	if self:Me(args.destGUID) then
 		self:OpenProximity(65121, 10)
-		self:Say(65121)
+		self:Say(65121, nil, nil, "Searing Light")
 	end
 	self:TargetMessageOld(65121, args.destName, "red", "alert")
 	self:TargetBar(65121, 9, args.destName, L.lightbomb_other)