+++ BigWigs/CHANGELOG.md
@@ -1,7 +1,15 @@
 # BigWigs
 
-## [v350.1](https://github.com/BigWigsMods/BigWigs/tree/v350.1) (2024-07-31)
-[Full Changelog](https://github.com/BigWigsMods/BigWigs/compare/v350...v350.1) [Previous Releases](https://github.com/BigWigsMods/BigWigs/releases)
+## [v350.2](https://github.com/BigWigsMods/BigWigs/tree/v350.2) (2024-08-02)
+[Full Changelog](https://github.com/BigWigsMods/BigWigs/compare/v350.1...v350.2) [Previous Releases](https://github.com/BigWigsMods/BigWigs/releases)
 
-- Core/BossPrototype: Attempt to fix a rare bad target scan  
-- Update deDE (#1728)  
+- Plugins/Bars: Another maxTime logic improvement  
+- Plugins/Bars: Randomize the test bar icons  
+- Plugins/Bars: Improve maxTime logic for bars that restart  
+- Remove the temporary bar reset message  
+- Change how we store some stats  
+- Update zhCN (#1730)  
+- VaultOfTheIncarnates/RaszagethTheStormEater: Add more heroic stage 3 timers  
+- Update deDE (#1729)  
+- Loader: Add support for content addons defining a minimum required version of the core  
+- Core/BossPrototype: Fix whitespace