+++ BigWigs_WrathOfTheLichKing/CHANGELOG.md
@@ -1,7 +1,9 @@
 # BigWigs [Wrath of the Lich King]
 
-## [v10.2.15](https://github.com/BigWigsMods/BigWigs_WrathOfTheLichKing/tree/v10.2.15) (2024-05-08)
-[Full Changelog](https://github.com/BigWigsMods/BigWigs_WrathOfTheLichKing/compare/v10.2.14...v10.2.15) [Previous Releases](https://github.com/BigWigsMods/BigWigs_WrathOfTheLichKing/releases)
+## [v11.0.0](https://github.com/BigWigsMods/BigWigs_WrathOfTheLichKing/tree/v11.0.0) (2024-08-05)
+[Full Changelog](https://github.com/BigWigsMods/BigWigs_WrathOfTheLichKing/compare/v10.2.15...v11.0.0) [Previous Releases](https://github.com/BigWigsMods/BigWigs_WrathOfTheLichKing/releases)
 
+- Fix `:Say` usage  
+- Revert "Update encounter IDs"  
+- Define X-BigWigs-Minimum  
 - bump toc  
-- Citadel/LichKing: Fix stuck infest bar